pyproject.toml
[project]
name = "financial-analysis-agent"
version = "0.1.0"
description = "A2A Agent for generating personalized A2UI financial analysis"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "google-genai>=1.0.0",
    "google-cloud-storage>=2.10.0",
    "python-dotenv>=1.0.0",
    "certifi>=2023.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
